Move examples to another file & Make main filename snake case
